fix : delete book (admin)

+++ src/main/java/com/example/library/service/BookService.java
@@ -1,7 +1,9 @@
 package com.example.library.service;
 
+import com.example.library.repository.BorrowHistoryRepository;
 import com.example.library.model.Book;
 import com.example.library.repository.BookRepository;
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import java.util.List;
@@ -11,6 +13,8 @@ public class BookService {
 
     @Autowired
     private BookRepository bookRepository;
+    @Autowired
+    private BorrowHistoryRepository borrowHistoryRepository;
 
     public List<Book> getAllBooks() {
         return bookRepository.findAll();
@@ -20,8 +24,15 @@ public class BookService {
         return bookRepository.save(book);
     }
 
-    public void deleteBook(int id) {
+    @Transactional
+    public String deleteBook(int id) {
+        if (!bookRepository.existsById(id)) {
+            return "Book not found";
+        }
+
+        borrowHistoryRepository.deleteByBookId(id);
         bookRepository.deleteById(id);
+        return "Book deleted successfully";
     }
 
     public List<Book> searchBooks(String keyword) {
